requirements.txt
arrow==1.2.3
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
-e git+https://github.com/aybidi/cli-telemetry.git@39246e10b09c91be07846fbe8cdf4efe11b84fe0#egg=filesys
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==6.0
requests==2.28.1
six==1.16.0
text-unidecode==1.3
urllib3==1.26.12
